refactor: focus & mood 케이스를 diary 하나로 통일
- 들어가는 이미지에 titleLabel을 없애냐 안 없애냐 차이
- 케이스를 별도로 분리하지 않고 하나로 통일
