+++ toduck/Source/Design/Component/CalendarView/Calendar/CalendarType.swift
@@ -8,6 +8,5 @@
 enum CalendarType {
     case toduck // GNB - 좌측
     case sheet  // 일정&루틴 - 날짜
-    case mood   // 일기 - 기분
-    case focus  // 일기 - 집중도
+    case diary  // 일기 - 기분 & 집중도
 }
